fix up mode checking, fix up file_write
mode should return True or False, not None because this value can make
its way back into a call to mode_* which will not always have the
intended effect.

change file_write so that the "sudo" parameter overrides whatever
context it may be executing in.

fix call to mode_sudo that was using "sudo" instead of "use_sudo" in
file_write.