Silence apt-get progress meter with -q (--quiet)
In normal usage, `apt-get` displays % progress meters on stdout. But
these meters are intended for terminal output not logging output.

With the fabric pseudo-terminal (`pty=True`), the meters do not display
as intended and instead produce a newline every time there is any
progress.

This commit adds the `-q` option to all the package_*_apt() functions
that call `apt-get`.